requirements-pro.txt
# Requirements for Pro Tier
# 专业版依赖 - 深度学习 + 持续进化

# Include all Plus tier dependencies
pymysql>=1.0.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
statsmodels>=0.14.0
scikit-learn>=1.0.0
pyod>=1.1.0
shap>=0.44.0

# Pro tier additions - Deep learning & causal inference
torch>=2.0.0
pytorch-forecasting>=1.0.0
dowhy>=0.10.0
